+++ packages/core/src/drivers/db/models/index.js
@@ -1,3 +1,4 @@
+// SIIGES 2.0
 const { Ciclo, CicloSchema } = require('./ciclo');
 const { Domicilio, DomicilioSchema } = require('./domicilio');
 const { Diligencia, DiligenciaSchema } = require('./diligencia');
@@ -119,6 +120,12 @@ const { AsignaturaEquivalentePrograma, AsignaturaEquivalenteProgramaSchema } = r
 const { FundamentoServicioSocial, FundamentoServicioSocialSchema } = require('./fundamentoServicioSocial');
 const { TipoModalidad, TipoModalidadSchema } = require('./tipoModalidad');
 const { ModalidadTitulacion, ModalidadTitulacionSchema } = require('./modalidadTitulacion');
+const { SolicitudBeca, SolicitudBecaSchema } = require('./solicitudBeca');
+const { SolicitudBecaAlumno, SolicitudBecaAlumnoSchema } = require('./solicitudBecaAlumno');
+const { TipoSolicitudBeca, TipoSolicitudBecaSchema } = require('./tipoSolicitudBeca');
+const { EstatusAlumnoBeca, EstatusAlumnoBecaSchema } = require('./estatusAlumnoBeca');
+const { EstatusSolicitudBeca, EstatusSolicitudBecaSchema } = require('./estatusSolicitudBeca');
+
 // Siiges 1.0
 const { Academia, AcademiaSchema } = require('./academias');
 const { AlumnoObservacion, AlumnoObservacionSchema } = require('./alumnoObservaciones');
@@ -156,12 +163,6 @@ const { SolicitudUsuario, SolicitudUsuarioSchema } = require('./solicitudes_usua
 const { Testigo, TestigoSchema } = require('./testigos');
 const { TituloElectronico, TituloElectronicoSchema } = require('./titulosElectronicos');
 // Siiges 1.0 Fin
-
-const { SolicitudBeca,SolicitudBecaSchema} = require('./solicitudesBecas')
-const { SolicitudBecaAlumno, SolicitudBecaAlumnoSchema} = require('./solicitudesBecasAlumnos')
-const { TipoSolicitudBeca, TipoSolicitudBecaSchema} = require('./tipoSolicitudesBecas')
-const { EstatusAlumnoBeca, EstatusAlumnoBecaSchema} = require('./estatusAlumnosBecas')
-const { EstatusSolicitudBeca,  EstatusSolicitudBecaSchema} = require('./estatusSolicitudesBecas')
 
 function setupModels(sequelize) {
   // Initialize models
@@ -530,11 +531,11 @@ function setupModels(sequelize) {
   TituloElectronico.associate(sequelize.models);
   // Siiges 1.0 Fin
 
-  SolicitudBeca.init( SolicitudBecaSchema, SolicitudBeca.config(sequelize));
+  SolicitudBeca.init(SolicitudBecaSchema, SolicitudBeca.config(sequelize));
   SolicitudBecaAlumno.init(SolicitudBecaAlumnoSchema, SolicitudBecaAlumno.config(sequelize));
-  TipoSolicitudBeca.init( TipoSolicitudBecaSchema,TipoSolicitudBeca.config(sequelize));
-  EstatusAlumnoBeca.init( EstatusAlumnoBecaSchema,EstatusAlumnoBeca.config(sequelize));
-  EstatusSolicitudBeca.init( EstatusSolicitudBecaSchema,EstatusSolicitudBeca.config(sequelize));
+  TipoSolicitudBeca.init(TipoSolicitudBecaSchema, TipoSolicitudBeca.config(sequelize));
+  EstatusAlumnoBeca.init(EstatusAlumnoBecaSchema, EstatusAlumnoBeca.config(sequelize));
+  EstatusSolicitudBeca.init(EstatusSolicitudBecaSchema, EstatusSolicitudBeca.config(sequelize));
 }
 
 module.exports = setupModels;
